fixed script and updated text
Replaced hipster ipsum text.
Each image has its own click event in the script.  This seems inefficient and doesn't go with the dry principle but for now at least it works and I can attempt to refactor  later. Also to css I added a z-index of 1 to the nav so it lays overtop of the images on mobile view.   Can now do all of this on the other two pages.